requirements.txt
streamlit>=1.0.0
pandas>=1.5.0
matplotlib>=3.5.0
beautifulsoup4>=4.10.0
selenium>=4.0.0
requests>=2.28.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
lxml>=4.9.0
pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.5.0
